[herd] Add possibility for Arch specific binop
This commit also fixes a problem for Java architecture, which was not
build with herd, and never called.

This commit adds a variant to `Op.op` and it now takes a type argument
as `Op.op1`.

To make it easier, I've added a functor in `ArchOp` to take previous
`ArchOp` and make them compatible. The types `no_arch_op` and
`no_arch_op1` are phantom types and common to all. They are used by the
AST of C and Java.

It is now possible to add new Arch-dependent binary operations, e.g.
bitvector concatenation in ASL.