[all] Some extensions of the X86_64 architecture
[lib]: Separate x86_64 tokens into check_name function

Motivation: We wish to check the compilation of x86_64 binaries, which
require us to be able to parse x64 ELF binaries.

As such we need to export X64 tokens for use in other parsers.

This patch updates the X64 grammar so it is in line with the AArch64
grammar.

[test]: ADD X86_64 testing

Motivation: We wish to test the compilation of x86 64 programs.

To do this we need to add x86 64 instructions, these instructions also
need tests.

This commit adds the necessary changes to test x86_64 programs. See for
example A000.litmus

[all]: Add Ret instruction support

Motivation: We wish to test the compilation of X86_64 programs.

Compilers emit RET, thus this patch adds support for the RET/RETQ
instruction.

 Consider the following test emitted by gcc -Og:
 ```
 X86_64 S00009

 { [P1_r0]=0;[x]=0;[y]=0; }
 (*****************************************************************)
 (* Compiler:                                                     *)
 (* x86_64-linux-gnux32-gcc-10 -c -g -Og -pthread --std=c11 -march=x86-64*)
 (*****************************************************************)
   P0            |  P1                 ;
    movl $2,(x)  |   mov (y),%eax      ;
    movl $1,(y)  |   cmp $1,%eax       ;
    ret          |   je L0x3b          ;
                 |  L0x31:             ;
                 |   mov %eax,(P1_r0)  ;
                 |   ret               ;
                 |  L0x3b:             ;
                 |   movl $1,(x)       ;
                 |   jmp L0x31         ;

 exists ([x]=2 /\ P1_r0=1)

 ```
 To support this test we must implement RETs. This patch adds support
 for this instruction

[lib]: X86_64 parse RIP register

Motivation: We wish to test the compilation of x86_64 registers

To do this we need to parse the rip register.

Parsing needs testing.

This patch parses the RIP register and adds tests in herd as appropriate

[lib]: X86_64 add nopl 0x0(%rx,%rd,k) support

Motivation: We wish to test the compilation of x86 64 programs.

Compilers emit nopl/nopw instructions of the above form.

As such we need to parse and support this syntax.

Tests are added that regard the above expressions as NOPS.

[lib]: x86_64 Added CS register

Motivation: we wish to test the compilation of x86_64 code

To do this we need to add the CS register.

This requires parsing. This patch adds parsing support for the CS
register.

Tests are also added in Herd

[lib]: X86_64 Parsing support for %cs:0x0(%rd,%rs,k) syntax

Motivation: We wish to test the compilation of X86 64 programs.

Compilers emit `NOPL %cs:0(%rax,%eax,0);`

We must parse this syntax, this patch adds support for that.

Tests are also provided in herd

[lib]: X86_64 endbr64 as NOP

Motivation: We wish to test the compilation of X86_64 code

To do this we need to parse the endbr4 instruction (which is a nop)

This patch adds parsing support for endbr64 as a NOP.

Tests are also provided for herd.

[lib]: X86_64 Nop Parsing

Motivation: We wish to test the compilation of x86 64 instructions.

To do this we need to parse the NOP instruction.

This patch adds support for NOP.

A test was added `A000` in a previous patch

[all]: Added X86_64 AND Instruction

[all]: Add X86_64 SHL instruction

[herd] guard ret on Telechat Variant, merge AFTER PPC patches

[herd] Add test for scaled indexed access