[herd] Add variant lse2 which implements the Armv8-A FEAT_LSE2
The new lse2 variant changes the default semantics for pair
instructions and ensure that their explicit belong to the same
atomicity class.

Note that for now, herd doesn't perform any checks on the alignment of
the data locations accessed by the pair instruction.

In addition, it changes the definition of the sca-class relation, to
exclude the id relation. This change doesn't make any changes to the
model as all uses of sca-class are optional and thefore the id subset
of the old sca-class relation is unnecessary.

Signed-off-by: Nikos Nikoleris <nikos.nikoleris@arm.com>