[herd] New semantics for CAS
This commit edits the intra-instruction dependencies from the
CAS semantics, on the case where the comparaison is successful
(ok case).
Changes:
a. the iico_data to the Write Register Event, that was previously
   starting from the Read Register Event of the value to be compared, is
   now starting from in two new executions either the Read Register
   Event of the value to be compared, either from the Read Memory Event.
b. the iico_ctrl from the Read Register Event of the value to be
   compared to the Write Register Event is now deleted.
c. the iico_ctrl from the Read Memory Event to the Write Register Event
   is now only present on the execution where the iico_data in a. is
   comming from the Read Register Event.

To justify those changes, we add litmus tests to test those
dependencies:
1. MP+rel+CAS-ok-RsRs-addr the existence of an execution where there is
   no iico_data from the Read Register Event of the value to be compared
   to the Write Register Event.
2. MP+rel+CAS-ok-MRs-addr tests the existence of an execution where
   there is no iico_data from the Read Memory Event to the Write
   Register Event.
3. MP+rel+CAS-ok-bothRs-addr tests the absence of any execution where
   there is neither the iico_data from the Read Register Event of the
   value to be compared to the Write Register Event nor the iico_data
   from the Read Memory Event to the Write Register Event.
4. LB+rel+CAS-ok-RsRs-addr tests the existence of an execution where
   there is no iico_ctrl from the the Register Read Event of the value
   to be compared to the Write Register Event.
5. LB+rel+CAS-ok-MRs-addr tests the presence of an iico_ctrl between the
   Memory Read Event to the Write Register Event.

A lof of tests have impacted logs, because of the number of execution
being changed by this. No other test in the catalogue change of kind.