[herd,asl] Implement repetitive statements
The three repetitive statements, while, repeat and for loops, are
implemented. Loops that are not known to be bounded are unrolled
up to _u_ times. The value of _u_ is given as an "unroll' parameter,
on the command-line (`-unroll <int>`) or in a configuration file.
The default value is higher for ASL.