[herd] Refactor DC and IC actions into a common CMO
This patches creates an architecture independent CMO action and uses
it to encapsulate the DC and IC events (for AArch64). In parallel, it
cleans up the relevant eventsets for AArch64, and similarly to what
herd does for barriers, every DC and IC instruction will raise an
eponymous event (a dc.cvau,Xt will raise a DC.CVAU event, etc).

Signed-off-by: Nikos Nikoleris <nikos.nikoleris@arm.com>