[asl] Three register ADD and SUB, plus ADDS and SUBS
To implement "S" variants of operations, _i.e._ operations
that set NZCV flags, we implement AArch64 PSTATE.

The implementation is quite tricky:
  + This state register has a double personality:
    PSTATE for ASL, NZCV for AArch64.
  + Moreover PSTATE is a bitvector and  NZCV in an integer.
  + NZCV is read from context, so has to be PSTATE.

Finally a solution is as follows: events are emitted with NZCV as a
location and an integer as value, while the interpreter handle PSTATE
as a global register of the appropriate bitvector type. The most
tricky part is the initialisation of PSTATE from context.