[lib]: Add ARM Push/POP and FP Register
Motivation: We wish to test compilationt targeting Armv7 programs.

To do so we must parse PUSH/POP operations and the FP register.

This patch adds support for these instructions/registers.

Since we do not handle program prolog/epilogs or SP/FP relative
accesses, PUSH and POP are NOPs - however we still need to parse
them in real code (and remove as appropriate).