[all] Additional PPC instruction implementations
[all]: Add Testing for PPC instructions

[all]: Add PPC LIS instruction

[all]: ADD PPC HWSYNC instruction

[all]: Add rlwinm, rlwimi and clrldi PPC instructions

Motivation: We wish to test the compilation of PPC programs

Improvements:
  + Replace left shifts by rotate
  + Use binary representation directly as the argument of `Op.AndK`.

Co-Authored-By: Luc Maranget <Luc.Maranget@inria.fr>

[all]: Added cmplwi PPC instruction

motivation: we wish to test the compilation of PPC instructions

To do so we need to parse the cmplwi instruction.

This patch adds support for this type

[all]: Add PPC ADDIS instruction

Motivation: We wish to test the compilation of programs targeting Power
PC.

To do so, we must parse the ADDIS instruction emitted by GCC.

ADDIS is add immediate, shifted.

[all]: Add LWA instruction

Motivation: We wish to test the compilation of programs targeting
PowerPC.

To do so we must parse LWA (load word algebraic) instructions.

Instruction lwa performs sign extension, while
instruction lwz performs zero extension.

[lib]: Add PPC Nop instruction

[all]: add extsw PPC instruction

Motivation: We wish to test the compilation, targeting PPC

To do so we must test the extsw instruction.

This patch adds support for this instruction

[herd]: Fix blr definition in PPC

Consider the following litmus test - output by GCC:
```
PPC LB03711_avoid_fenced_int_C_tests

{ [P0_r0]=0;[P1_r0]=0;[x]=0;[y]=0;uint64_t %P0_P0_r0=P0_r0;uint64_t %P0_x=x;uint64_t %P0_y=y;uint64_t %P1_P1_r0=P1_r0;uint64_t %P1_x=x;uint64_t %P1_y=y }
(*****************************************************************)
(* Compiler:                                                     *)
(* powerpc64le-linux-gnu-gcc-10 -c -g -O1 -pthread --std=c11 -ffreestanding -mabi=elfv1 -fno-section-anchors*)
(*****************************************************************)
  P0                    |  P1                     ;
   hwsync               |   hwsync                ;
   lwz r9,0(%P0_x)      |   lwz r9,0(%P1_y)       ;
   cmpw r9,r9           |   cmpw r9,r9            ;
   b   L0x18            |   b   L0x6c             ;
  L0x18: isync          |  L0x6c: isync           ;
   clrldi r9,r9,32      |   extsw r10, r9         ;
   lwsync               |   cmpwi r9,1            ;
   hwsync               |   beq  L0x8c            ;
   li r8,1              |  L0x7c:                 ;
   stw r8,0(%P0_y)      |   stw r10,0(%P1_P1_r0)  ;
   stw r9,0(%P0_P0_r0)  |   blr                   ;
                        |  L0x8c:                 ;
                        |   li r8,1               ;
                        |   stw r8,0(%P1_x)       ;
                        |   b   L0x7c             ;

exists (P0_r0=1 /\ P1_r0=1)
```
The blr instruction 'exits' from the program -  to support this test
we need to support blr

This patch interprets the `blr` instruction as `jump to end of code`
(or exit) under the control of the new `telechat` variant.

When the variant is not specified, a fatal error is flagged.

[all]: Add PPC lwax instruction.

Motivation: We wish to test the compilation of gcc on targeting PowerPC.

To do so we must add the lwax instruction.

This patch adds support for this instruction