[herd] Immaterial change in SWP
This commit removes the iico_order dependency between the two register
events of a SWP instruction. As this dependency was never used by the
model after being constructed, it will not have any consequence on any
test, and this change is immaterial.

Idem for the atomic operations instruction, such as LDADD, STEOR, etc.

The test LB+SWP-RsRt-addr+rel.litmus tests for the absence of any
intra-instruction data or control dependencies between the two register
events of the SWP instruction.