[herd,asl] Fix transmission of data ports from ASL to AArch64
The fix is simple: the showing of event set `AArch64_DATA`
was missing in file `asl.cat`.

Notice that "showing" sets and relations is a flexible
mean to transmit them from ASL to AArch64.

Also notice that `asl.cat` already showed the identity _relation_
over the same event set as `AArch64_DATA`. We leave it,
as it apparently does not harm.