[herd] Fix AArch64Sem lift_memop for -variant vmsa,memtag
This change modifies the semantics of the translation (-variant vmsa)
for checked memory instructions when MTE is enabled (-variant
memtag). In short, for a checked memory instruction that performs a
tag check prior to the explicit memory event, it allows the tag check
operation to have its own translation (read of the PTE followed by a
branching decision), separate from the translation that happens and
leads to the explicit memory event.

Also it adds support for a new type of memory, TaggedNormal featured
in FEAT_MTE2. TaggedNormal is the only type for which we peform a tag
check. To illustrate this with an example, in the following test, the
STR is a checked access and will result in a TagCheck fault.

AArch64 STR-TaggedNormal-red
{
 [PTE(x)]=(oa:PA(x),attrs:(TaggedNormal));
 0:X0=x:red;
}
 P0          ;
 MOV W1,#1   ;
L0:          ;
 STR W1,[X0] ;
locations[fault(P0:L0,x,TagCheck);]
exists(x=1)

While the following test, the STR will not raise a TagCheck fault:

AArch64 STR-Normal-red
{
 [PTE(x)]=(oa:PA(x),attrs:(Normal));
 0:X0=x:red;
}
 P0          ;
 MOV W1,#1   ;
L0:          ;
 STR W1,[X0] ;
locations[fault(P0:L0,x,TagCheck);]
exists(x=1)

Signed-off-by: Nikos Nikoleris <nikos.nikoleris@arm.com>