[litmus] Limit pointer to code
At the moment pointer to code is possible only
for architecture with fixed-size instructions.
For those, the size of instructions is defined
in the file `instruction.h`.

Hence consider that no pointer to code can be
saved, or handled in any way when the file
`instruction.h` does not exist.