[herd] Add a debugT monadic operation
This change adds debugT which can be used to print the basic fields of
the event structure in the provided monad type as its argument.

For example one can do

let m = M.debugT "m: " m in

which will leave m unmodified and print the event_structure of m in
stderr.

Signed-off-by: Nikos Nikoleris <nikos.nikoleris@arm.com>