[herd] Restore architectural intent for AArch64 SWP/CAS that return to ZR
Much like LD<op>, SWP and CAS instructions that return to ZR (e.g.,
WZR), are not regarding as doing a read for the purposes of a
DMB.LD. In addition, such instructions do not load with acquire
semantics. This commit restores the intent in the implementation of
the semantics of SWP and CAS in herd7 following discussions with the
architects.

In addition, this commit adds tests to the catalogue that demonstrate
the semantics of LDADD/SWP/CAS when they to WZR and when they return
to any other general purpose register.

Signed-off-by: Nikos Nikoleris <nikos.nikoleris@arm.com>