[AArch64 cat] Revamp definition of same-loc relation
1. Two Memory Effects E1 and E2 are related with same-loc, if the E1
is to the same location as E2.

2. If either E1 or E2 is an MMU Fault Effect where the Output Address
is known (AccessFlag or Permission Fault) then E1 and E2 are related
with same-loc if the low-order-bits of the address associated with the
Fault Effect are the same as the low-order-bits of the Memory Effect
and the associated Implicit TTD Reads have the same Output Address.

3. If either E1 or E2 is an MMU Fault Effect where the Output Address
is not known (Translation Fault), then E1 and E2 are related with
same-loc, if the low-order-bits of the address associated with the
Fault Effect are the same as the low-order-bits of the Memory Effect.

Signed-off-by: Nikos Nikoleris <nikos.nikoleris@arm.com>