[catalogue] Add support for running tests in aarch64-{VMSA,ETS2}
This change adds Makefile support for running the tests in
catalogue/aarch64-VMSA/shelf.py with the VMSA and ETS2
configurations. The way we run test also checks their outcome against
their respective kinds file. In addition, this change modifies
catalogue/aarch64-VMSA/shelf.py to include almost all tests (except
for some very large ones) and adds some missing tests in the kinds.txt
for VMSA and ETS2. Note that because the VMSA and ETS2 catalogues are
fairly large the new test targets are not include in the default test.

To run the new tests:

$> make vmsa-tests
$> make ets2-tests

Signed-off-by: Nikos Nikoleris <nikos.nikoleris@arm.com>