[Makefile] Add a test-all target
The Makefile used to offer an option to run all tests. As we added
more tests, the overhead for running all tests becomes
non-trivial. More notably, the vmsa-test and ets2-test targets include
more litmus tests than any other. In addition, the fact that make test
is invoked in the precommit hook makes it crucial that make test
terminates as quickly as possible. This commit adds another target
test-all which is inclusive of all tests.

Signed-off-by: Nikos Nikoleris <nikos.nikoleris@arm.com>