[all] Split precision into MTE precision and Fault handling
Prior to this patch, the tools were using a unified concept of
Precision which was used to decide how to act on an MTE tag mismatch
(raise an exception or set the TFSR_ELx) and how to handle a fault
(exception). This patch splits the precision into two different
arguments, the MTE precision and the fault handling mode.

Signed-off-by: Nikos Nikoleris <nikos.nikoleris@arm.com>