[herd,aarch64,asl] Add branching effects
Branching effects are added by the choice operator.

Some of them are passed to AArch64. Selected effects
are the ones with that are reachable from non-branching
effects and from which  non-branching effects are reachable.
Here, "reachable" means "in relation by the transitive closure
of (ASL) iico_control.

Branching effects of nature `Bcc` are identified
as the sources of iico_control relations to `PC`
write effects. Others branching effects are of nature `Pred`.

[hers,asl] ASL computation of branch conditional effects