[all] Implement INTID locations and corresponding values
This commit adds a new type of symbolic system location, representing
interrupt locations, and the support for the type of values that this
location can hold. This is the first step in adding support for GICv5
to Herd.

As per the spec, an INTID location can hold a value of the following
form: [INTID(x)]=(pending:0, enabled:1, priority:0, target:P0,
target_mode:<targeted>, trigger_mode=<edge>) where the values after
the colon are the corresponding default values for each field. An
interrupt can be pending or idle, enabled or disabled, it has a
priority represented by an integer (the higher the number, the more
priority is given to the interrupt), a target, represented by a
process number, a target mode, which can be targeted or 1ofN, and a
trigger mode, chosen between edge-triggered and level-sensitive. For
more information, visit the Arm GICv5 manual.  A generic interface,
along with concrete implementations, are added for intid values. The
scope of this change is enabling herd to parse intid locations as part
of the init state, and check their existence in the
postcondition. Support for GICv5 instructions will follow.