[herd] Fix fault behaviour for -variant vmsa,mte
This change simplifies the way faults are generated and handled when
variants vmsa and mte are both enabled at the same time. On a high
level, a checked memory instruction will generate two separate ops:

1) The tag op that will perform a translation, followed by a tag read
and a tag check if the translation permits it (attrs:(TaggedNormal)), and

2) The data op that is used to translate for the data access and check
if its necessary to wait for tag op (attrs:(TaggedNormal)).

These two perform separate translations and they generate separate and
independent Implicit PTE Memory Read effects.

Signed-off-by: Nikos Nikoleris <nikos.nikoleris@arm.com>