[all] Implement IntidUpdateVals
This commit adds the capability for herd to parse an IntidUpdateVal tuple,
which is of the form (intid: <intid_loc>, [<field_name>: <field_val>]*).
This new type of value is to be used by GIC instructions intending
to update fields of an interrupt location. As an example, the (yet
to be implemented) GIC CDPRI updates the priority field of an
interrupt location and is used as follows:

AArch64 gic.cdpri
{
 [INTID(A)]=(priority:0);
 X1=(intid:A, priority:1);
}
GIC CDPRI, X1

This updates the priority field of interrupt location A from 0 to 1.
The support for the various GIC instructions and their corresponding
effects will follow in a later commit.