[herd] Set ELR_EL1 automatically when it is possible
This change enables herd to handle tests like:

AArch64 LDRv0+I2V-dsb.ishst
{
 [PTE(x)]=(oa:PA(x),valid:0);
 x=1;

 0:X1=x;
 0:X3=PTE(x);
 0:X2=(oa:PA(x),valid:1);
}
 P0            | P0.F           ;
L0:            | STR X2,[X3]    ;
 LDR W0,[X1]   | DSB ISHST      ;
               | ERET           ;
locations[0:ELR_EL1;]
forall(0:X0=1)

Note that in the above test ELR_EL1 is automatically set to
P0:L0. However this wouldn't be the case if the faulting instruction
didn't have a label. In which case herd7 would warn that the target of
ERET is unknown.

Signed-off-by: Nikos Nikoleris <nikos.nikoleris@arm.com>