[asl] Hide stdlib filename from error messages.
This commit is premilinary work to adding Unreachable to the stdlib.
It alows errors to be thrown in the stdlib without its location being
disclosed in the error message.

This way tests can trigger exceptions in the stdlib without being
dependent on where the stdlib is.

Note that if the flag `_debug` in `builder.ml` is set, the actual file
is shown.