[catalogue] Fix litmus7 warnings for aarch64-mixed
litmus7 and the compiler used for the source code produced by litmus7
emits warnings for some of the tests in aarch64-mixed catalogue.

In most cases, the type of the register is declared incorrectly (for
example, uint64_t instead of uint32_t). In other cases, we need to
trancate a register from 64bit to 32bit.

This commit fix these warnings. There should be no change to the
semantics of any of these tests and for this reason their hash is
changed to match that of the previous version of the tests.

There is one warning which is benign:

Warning: File "./tests/LB+dmb.sy+data-wsi-wsi+MIXED.litmus" Register x0 has different types: <int64_t> and <int>

In this case, litmus7 is being a little pedantic. gcc seems to be fine
with its code.

Signed-off-by: Nikos Nikoleris <nikos.nikoleris@arm.com>