[herd] Add support for BPF Architecture
This commit adds the Lexer and the Parser for BPF assembly and also
implements all required files to support running litmus tests using herd.

Currently only herd is supported and support for litmus, etc. will be
added later.

The newly created files are written in the janestreet ocaml format and I
would like to enforce that even when these files are modified in future.

This implements fixed size support for the BPF instructions. The
following instructions are supported:

1. Arithemetic and Logical Instructions (+,-,*,/,%,&,|,^,<<,>>)
  a. Register to Register
    r0 += r1
  b. Immediate to Register
    r3 *= 44

2. Load / Store Instructions
  a. Load Register
    r0 = *(u8 *)(r0 + 0)
    r2 = *(u16 *)(r4 + 0)
    r4 = *(u32 *)(r5 + 0)
    r5 = *(u64 *)(r6 + 0)
  b. Store Register
    *(u8 *)(r0 + 0)  =  r0
    *(u16 *)(r4 + 0) =  r2
    *(u32 *)(r5 + 0) =  r4
    *(u64 *)(r6 + 0) =  r5
  c. Store Immediate
    *(u8 *)(r0 + 0)  = 23
    *(u16 *)(r4 + 0) = -32
    *(u32 *)(r5 + 0) = 44
    *(u64 *)(r6 + 0) = 55

3. Data movement instructions
  a. Register to Register
    r0 = r5
  b. Immediate to Register
    r3 = -40

4. atomic operations:
  a. 32 and 64 bit atomic ALU ops(+,|,&,^):
        lock *(u32 *)(rd + offset16) += rs
        lock *(u64 *)(rd + offset16) |= rs

  b. 32 and 64 bit atomic ALU ops with fetch:
        rs = atomic_fetch_add ((u64/u32 *)(rd + offset16), rs)
        rs = atomic_fetch_and ((u64/u32 *)(rd + offset16), rs)
        rs = atomic_fetch_or ((u64/u32 *)(rd + offset16), rs)
        rs = atomic_fetch_xor ((u64/32 *)(rd + offset16), rs)

  c. 32 and 64 bit atomic exchange:
        rs = xchg_64 (rd + offset16, rs)
        rs = xchg32_32 (rd + offset16, rs)

  d. 32 and 64 bit atomic compare and exchange:
        r0 = cmpxchg_64 (rd + offset16, r0, rs)
        r0 = cmpxchg32_32 (rd + offset16, r0, rs)

5. Jump instructions
  a. Unconditional jump
        goto lbl

  b. Conditional jump with register arguments
        if rs cond rd goto lbl  ; where cond is one of { ==, !=, <, >=}

  c. Conditional jump with one register and immediate argument.
        if rs cond imm32 goto lbl  ; where cond is one of { ==, !=, <, >=}

BPF supports 32 bit registers as well but they are not supported yet as
this commit doesn't add support for mixed size semantics.

A cat file has been added that defines the memory model of the BPF
architecture. This memory model is still under development and will be
updated as and when required.

Signed-off-by: Puranjay Mohan <puranjay@kernel.org>