[all] Parse  '<i>:<label>' for labels as inital values.
[litmus] Output labels in outcomes as herd does

[litmus] Global location initialisation with labels, `-mode presi`

Small fix, as initialisation of register location
was already implemented.

[litmus] Exclude code label global location from stabilisation

Checking stabilisation for global locations that contain
code labels proved a bit complex, as those are recorderd in
histograms as small integers. Excluding them is an
easy workaround.

Signed-off-by: Nikos Nikoleris <nikos.nikoleris@arm.com>

[litmus] Restore specific output for code labels

[litmus] Reject dangerous align requests

Meta data `Align=<id>:<nbytes>` in fact commands
allocations of location `<id>` is of size `<nbytes>`.
Test is rejected when `<nbytes>` is less then
the actual size of `<id>`.

If `<id>` size is not determined, as is the case of pointers for now,
a non-fatal warning is flagged.