[litmus] Add support for speedcheck parameter for -mode presi
This change adds support for the speedcheck parameter for -mode
presi. This was already supported in -mode std. The user can provide
the parameter "+sc" which will force the exit to as soon as the
post-condition is observed.

So as to avoid deadlocks, introduce a 2 level procedure
to stop the experiment.

Any instance that reaches the stop condition sets the global
`stop_now` flag. This global flag is copied into an
instance specific `stop_now` flag after each test
by the thread number zero of this instance.

Then, the instance threads synchronise on an instance level
barrier before reading the instance flag and interrupting
the test loop, if the flag is set.

After returning from the test loop, _all_ threads synchronise on
a global barrier, before reading the global flag and
interrupting the experiment if the flag is set.