[herd] Configure SVE vector length
SVE vector length can be configured to size <n>
at **herd** start up with `-variant sve:<n>`.

Size <n>, expressed in bits is a multiple of 128
between 128 and 2048, as specified by documentation.
Using `-variant sve` selects the current
default size of 128 bits.

Notice that the the metadata `Variant=sve:<n>` works as
expected for **herd** but triggers a warning  for **litmus**,
reflecting that hardware vector size will not be changed.