[litmus]: Use 'Upl' constraint modifier for predicate register
Constraint modifier we currently use for predicate register is

 `Upa` - Any of the SVE predicate registers (P0 to P15)

That made a trick so far, yet with GCC 14.1.0 it seems that register
allocation happens top to bottom, yet most of SVE instructions we
support allow predicate register only in a range P0 to P7. So build
fails.

Let's work it around by using `Upl` constraint modifier for predicate
register:

  `Upl` - One of the low eight SVE predicate registers (P0 to P7)

Signed-off-by: Vladimir Murzin <vladimir.murzin@arm.com>