[herd] Fix implementation of AArch64 SME `ADD<H|V>A`.
The `ADD<H|V>A` instructions scan only
a subpart (a tile) of the `ZA` storage.

Thus, since non-scanned elements must stay as they are,
the initial value of the scanning process is the old value of
the `ZA` storage and not zero.

Code is also changed to avoid read-store of the cells
that are unchanged because of the predicate registers,