[herd]: Fix address generation for STG, LDG and friends (again)
Partially revert c00eaf38b63 ("[herd]: Fix address generation for STG,
LDG and friends").

Altough it is correct that address generation using immediate signed
offset scaled by the Tag granule, it only applies to encoding of the
instruction but immediate for assembler has different requirement:

```
<simm> Is the optional signed immediate offset, a multiple of 16
in the range -4096 to 4080, defaulting to 0 and encoded in the "imm9"
field.
```
and it is what assembler implements

```
Error: immediate value must be a multiple of 16 at operand 2 -- `stg X1,[X0,#1]'
Error: immediate value must be a multiple of 16 at operand 2 -- `ldg X3,[X0,#1]'
```

So roll back herd part and update MTE tests, also add extra check for
immediate and dedicated litmus test to exercise the check.

Signed-off-by: Vladimir Murzin <vladimir.murzin@arm.com>