[litmus] Emit label related code only when instruction type is defined
All label related code should depend on the existence of
the "instruction.h" file in the architecture specific
sub-directory of `litmus/libdir`. The "instruction.h" file
define the C type of instructions `ins_t`.

In other words absence of this file means that we give up
on implementing code addressing for a given architecture.