[herd] Enable Xor'ing equal symbolic addresses.
This addition is necessary for some tests that
xor addresses as part of address dependencies.
Those tests abound in  the MTE catalogue and
fixing the unsound xor bug leads to prohibitive
execution times (up to x30).

Hence we also introduce concurrent execution
and timeout for those tests to mitigate this
slowdown.