[herd] Fix dangling rf arrows from regs outside of finals
A bug was observed on the master branch that, for tests writing
to local registers, the execution graph would show rf relations
from writes to local registers, pointing to nothing, as if
those registers were part of the final state, even though they
are not. This commit addresses the issue and makes herd only
draw such arrows when needed (i.e. when said registers are
actually part of the final state)