[herd,aarch64] Liveness condition
The condition will reject executions for which there exists
some `fr` source in a pruned po that is not followed
by a `rf` target:

That is, given
  +---W
  |
 \ /
  R ----po---> TooFar

there does not exists a rf s.t.

  +---W--+
  |      |
 \ /    \ /
  R -po--R---> TooFar

This condition is meaningful only then the set `TooFar` is non-empty.
That is, only on pruned executions, when the `TooFar variant is
active (`-variant TooFar`).