[herd,aarch64] Liveness condition
The condition will reject executions for which there exists
some `fr` source in a pruned `po` that is not followed
by a `rf` target:

That is, given
  +-fr->W
  |
  |
  R ----po---> CutOff

there does not exists a rf s.t.

  +-fr->W-rf-+
  |          |
  |         \ /
  R ----po---R-----> CutOff

Notice that this "liveness" condition applies to writes that
are followed by a DSB effect (_i.e._ stores followed by a DSB
fence instruction).