[lib] Fix PteVal parsing and add discriminate IntidVals
This fixes the issue that PteVal parsing had with the
valid field, which was throwing a parser error whenever
included as part of a PteVal. In addition, this adds a
way to reliably discriminate between PteVals and IntidVals.
The type of an IntidVal is now inferred when the left_loc
of the assignment is an intid location. For a register
or any other global location, a type annotation intval_t
is required. The gicarg type annotation for IntidUpdateVals
has also been renamed to gicval_t.