[all]: Refactor AArch64 EOR vector instructions
Currently parsing for EOR instructions for both Neon and SVE rely on
`OP` token so it has to assert for everything but EOR. Moreover, EOR
instructions itself have constrains:

 - Neon EOR accepts only .8B or .16B registers
 - SVE EOR accepts only .D register

Keeping all these checks in a parser is not great, we better do it in
validation pass. So let's refactor EOR vector instructions and plug
required checks into instruction validator.

Signed-off-by: Vladimir Murzin <vladimir.murzin@arm.com>