[asl] Use `integer{}` syntax to inherit integer constraints
The assignment `let x : integer{} = UInt(bv);` inherits integer
constraints from the right-hand side, so that users do not need to spell
them out explicitly. This is supported within tuples too:
`let x : (integer{}, boolean) = ...;`

We must take care not to support this syntax elsewhere, so in general
annotating a type that contains this form will cause an error.