[litmus] New `-variant noinit` option.
The new switch effect is _not_ initialising  memory
locations in object C code. This is useful for timing
purposes, especially when arrays are involved.

Accordingly, the initial values of memory locations
are absent from test pretty-printing.

Notice that this switch is not implemented in `presi` mode:
combining the two modes yields a fatal error.