[tools] Implement the dot2desc tool
This tool takes an execution graph in the form of a dot file and
describes its effects and relations between them in English. It
produces a Markdown file with inlined LaTeX, which contains
commands defined in the catdefinitions.tex file. pandoc needs to
be used to interpret those commands and produce the final, pure
Markdown output.