[herd,WIP] Introduce basic support for GCS
Guarded Control Stack (GCS), an area of memory in which procedure
return addresses and exception return addresses are stored and
protected from modification.

When the processor executes a Branch with Link instruction, such as
BL, the return address is pushed onto the GCS as well as being written
into the Link Register (LR). On a procedure return, the latest stored
return address is popped from the GCS. The processor either compares
the popped value with the LR, or uses the popped value directly.

There are times when the software needs to make manual adjustments to
the control stack to enable this, the architecture provides
instructions for maintaining the GCS.

Implementation choices and known limitations:

  - GCS memory model is not addressed
  - No overshooting memory effect
  - Permission model is not addressed
  - No support for RVCHKEN=0
  - Following GCS instructions are implemented:
    + GCSPOPM
    + GCSPUSHM
    + GCSSTR
    + GCSSS1
    + GCSSS2

Signed-off-by: Vladimir Murzin <vladimir.murzin@arm.com>