[asl] Changes to mutable assignment left-hand sides
Enforce a syntax of permitted left-hand sides, and desugar this to the
underlying representation during parsing.

Also add a check to ensure the same underyling variable is not written
to multiple times in a tuple assignment, i.e. forbid any combination of
`(x, x.foo, x[slices], x[[idx]])` and so on.