[asl] Fix dynamic domain after enum labels
Dynamic domain was still using the old definition of enumeration labels:
enumeration labels were compiled to integers. After PR #1091,
enumeration labels are first class citizens, and so we can update the
definition of dynamic domains to use those. This results in a simpler
and more straight forward definition of dynamic domains.