[herd] Optimised evaluation by implicit transitive relations.
Add implicitly transitive relations to the cat interpreter.

Implicitly transitive relations are
subject to two main optimisations :
  + When the acyclic test argument is a disjunction
   `(..|r|..)` and that argument `r` is
   a relation in implicit transitive form, there is
   no need to apply transitive closure. Namely
   `acyclic (..|r|..)` and `acyclic (..|r+|..)` are
   equivalent.
  + Intersection with a standard relation provides
    a support for repetitively testing connectivity,
    without effectively computing transitive closures.

The implicitly transitive relations are implemented
as an ordinary relation `r`. Their value is `r+`

Context where transitive closure need not be performed
(such as a disjunction argument in an acyclicity clause)
is implemented as a new boolean parameter to the
evaluation function.

Co-authored-by: Hadrien Renaud <hadrien.renaud2@arm.com>