[litmus] Handle memory attribute changes gracefully
Initialisations of memory locations (for example, x) take place before
we intialise the corresponding PTE value (for example, PTE(x)). In
some cases, handling the intialision of the PTE values necesitates a
change of memory attributes. When this happens, the intialisation of
the memory location uses the default memory attributes (Normal,ISH,WB)
and that means that the next access to that memory location will
happen with mismatcheched memory attributes.

To avoid creating the conditions for mismatched memory attributes at
initialisation, this patches changes litmus7 to perform cache
maintenance when needed.

Signed-off-by: Nikos Nikoleris <nikos.nikoleris@arm.com>