[doc] Capturing convention about 'fault' predicate
When introducing the SVC instruction, as part of the following commit:
https://github.com/herd/herdtools7/commit/62b1036808b0e0ce4d6ad306ba55b1b635d1fa6c,
it became apparent that there is a significant difference between checking the
address or label of the instruction that caused the fault and the return
address of the fault. This distinction emerged since SVC handler returns to the
instruction after the SVC, unlike most instructions which return to the instruction
that caused the fault.