[dot2desc] Rename params and event tags and get rid of access sizes in dot files
This commit adds the ability to overwrite the dot file passed with
a graph where the architectural registers are replaced by the symbolic
parameters passed through the -instr flag. It also gets rid of access
sizes (eg. A quad register read effect of X0 normally appears as
R0:X0q; this gets rid of the q), and replaces the tags (the letter
appearing at the beginning of a node description - eg. the 'a' in
"a: R0:X0q") with an event identifier Ei, where i is the index of
the event in the topological order of nodes.