[herd,aarch64] Force integer encoding of booleans in NZCV computation
When computing the status flags NZCV, it is important
not to introduce actual booleans.

As a matter of fact, the code assumes the encoding
of false as 0 and true as 1. If booleans are encoded differently
the old code fails. In practice this untimely situation occurs
where AArch64Sem code is executed in AArch64+ASL mode, which
indeed occurs. The fix consists in introducing some explicit
boolean to integer operations `(fun b -> b ? 1 : 0)`, and to change
the readBit operation output.