[asl][menhir2bnfc] Added comparison tool for bnfc ASL1 parser
A new tool parser_cmp has been added to compare between an autogenerated
menhir parser coming from the bnfc tool and the ASL1 parser.
To build this tool you need to enable the build of the `bnfc_parser`
library inside `menhir2bnfc` - note that bnfc is required to build it.