[litmus,aarch64] ELO mode by vector table.
We now systematically change the vector table
of EL0 threads, by introducing an indirect jump
in the slot for EL0 synchronous interruptions.

During test the indirection point to the appropriate
fault handler. Just before leaving test code,
the indirection is changed to point to
the end of test code. This technique
permits returning to EL1 at test end
by using a `SVC` instruction.

Moreover, `SVC` instructions that may be present
in test code are treated by the appropriate fault handler.